# swagger.yaml
openapi: 3.0.1
info:
  title: Food Trucks Service API
  version: v1
paths:
  '/api/v1/FoodTrucks/closest/{latitude}/{longitude}':
    get:
      tags:
        - FoodTrucks
      parameters:
        - name: latitude
          in: path
          required: true
          schema:
            type: number
            format: double
        - name: longitude
          in: path
          required: true
          schema:
            type: number
            format: double
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FoodTruck'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '204':
          description: Success
  '/api/v1/FoodTrucks/closest/{latitude}/{longitude}/{milesRadius}':
    get:
      tags:
        - FoodTrucks
      parameters:
        - name: latitude
          in: path
          required: true
          schema:
            type: number
            format: double
        - name: longitude
          in: path
          required: true
          schema:
            type: number
            format: double
        - name: milesRadius
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FoodTruck'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '204':
          description: Success
components:
  schemas:
    FacilityType:
      enum:
        - None
        - Truck
        - PushCart
      type: string
    PermitStatus:
      enum:
        - Inactive
        - Requested
        - Expired
        - Suspend
        - OnHold
        - Approved
        - Issued
      type: string
    FoodTruck:
      type: object
      properties:
        locationId:
          type: integer
          format: int32
        applicant:
          type: string
          nullable: true
        facilityType:
          $ref: '#/components/schemas/FacilityType'
        locationDescription:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        permit:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/PermitStatus'
        foodItems:
          type: string
          nullable: true
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        schedule:
          type: string
          nullable: true
        daysHours:
          type: string
          nullable: true
      additionalProperties: false
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties:
        type: object
        additionalProperties: false